% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.dbicc.R
\name{boot.dbicc}
\alias{boot.dbicc}
\title{Bootstrap Confidence intervals for dbICC (internal)}
\usage{
boot.dbicc(dmat, nsub, nmea, bootsamp, adhoc = FALSE)
}
\arguments{
\item{dmat}{A distance matrix or an object of \code{dist}, of dimension \code{sum(nmea)*sum(nmea)}. 
Note that the structure of the distance matrix, with the rows or
columns is grouped by subjects or individuals. The details refer to Figure 1 of
Xu el at., 2020.}

\item{nsub}{Number of subject or individual.}

\item{nmea}{A vector containing number of the measurement for each subject or individual; 
if \code{nmea} is a scalar, it means each subject shares the same number of the measurement.}

\item{bootsamp}{A sample with replacement.}

\item{adhoc}{A logical variable, whether to apply the ad hoc correction when
estimating the dbICC from a bootstrap sample. Default is \code{FALSE}.}
}
\value{
A scalar, giving the dbICC value
}
\description{
Nonparametric bootstrapping can be used to construct confidence 
intervals for the Distance-based Intraclass Correlation Coefficient 
(dbICC) based on samples of the subjects with replacement.
}
\seealso{
\code{\link{plotdmat}},\code{\link{dm2icc}}
}
\author{
Meng Xu \email{mxu@campus.haifa.ac.il}, Philip Reiss
}
\keyword{internal}
