% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbicc-package.R
\docType{package}
\name{dbicc-package}
\alias{dbicc-package}
\title{Distance-based Intraclass Correlation Coefficient (dbICC)}
\description{
This package contains a series of internal and external functions of 
Distance-based Intraclass Correlation Coefficient (dbICC), which is a new way to 
access the reliability of new and complex types of data in terms of arbitrary 
distances among observations.
}
\references{
P.T. Reiss, M. Xu, I. Cribben (2019). Generalized reliability based on distances. To be submitted.
}
\seealso{
\code{\link{dm2icc}}
}
\author{
Meng Xu, Philip Reiss

Maintainer: Meng Xu <mxu@campus.haifa.ac.il>
}
