% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2icc.R
\name{dm2icc}
\alias{dm2icc}
\title{Distance-based Intraclass Correlation Coefficient (dbICC)}
\usage{
dm2icc(dmat, nsub, nmea)
}
\arguments{
\item{dmat}{a distance matrix or an object of \code{dist}, of dimension \code{sum(nmea)*sum(nmea)}. 
Note that the structure of the distance matrix, with the rows or
columns is grouped by subjects or individuals. The details refer to Figure 1 of
Xu el at. (2020).}

\item{nsub}{number of subject or individual.}

\item{nmea}{a vector containing number of measurement for each subject or individual; 
if \code{nmea} is a scalar, it means each subject shares the same number of the measurement.}
}
\value{
a scalar, giving the dbICC value
}
\description{
When the distance among any test-retest datatype can be defined, its reliability 
performed in Xu et al. (2020) can be obtained by this function, to calculate the 
value of dbICC through the related distance matrix.
}
\examples{

##Point estimates of dbICC

# Generation function for R^2 points from multi-normal distribution
R2gen<-function(nsub,nmea,m=1,variance,sds=NULL,pt=FALSE){
    if (is.null(sds)==FALSE) set.seed(sds)
    if (length(nmea)==1) nmea<-rep(1,nsub)*nmea
    sig1<-diag(rep(variance,2))
    mu<-c(0,0)
    sig2<-diag(rep(1,2)) # true-value variance of the 2-d normal distribution
    t<-MASS::mvrnorm(nsub,mu,sig2)
    e<-MASS::mvrnorm(sum(nmea),mu,sig1/m)
    p<-matrix(apply(t,2,rep,times=nmea),ncol=2)+e#(I*J)x2
    if (pt==TRUE) return(list(t=t,p=p))
    if (pt==FALSE) return(p)
}

# set the number of the point
I <- 10

# set the number of the measurement for each point
J <- 4

# generate the sample of R^2 points
varl <- .25 # error variance of the 2-d normal distribution
pij <- R2gen(I,J,variance=varl)

# calculate the squared distance matrix via Euclidean distance
distmat<-as.matrix(dist(pij))

#plot the distance matrix
plotdmat(distmat,I,J)

# dbICC value
dm2icc(distmat,I,J)


}
\references{
\itemize{
\item Xu, M., Reiss, P. T., and Cribben, I. (2020). Generalized reliability based on distances. Biometrics, to appear. \url{https://arxiv.org/abs/1912.07137}.
}
}
\seealso{
\code{\link{plotdmat}},\code{\link{dm2icc.bt}}
}
\author{
Meng Xu \email{mxu@campus.haifa.ac.il}, Philip Reiss
}
\keyword{dbICC,}
\keyword{reliability}
