% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdmat.R
\name{plotdmat}
\alias{plotdmat}
\title{Plot the distance matrix in dbICC}
\usage{
plotdmat(dmat, nsub, nmea, xlab = NULL, legend = TRUE)
}
\arguments{
\item{dmat}{See \code{\link{dm2icc}}.}

\item{nsub}{number of subject or individual.}

\item{nmea}{a vector containing number of  measurement for each subject or individual;
if \code{nmea} is a scalar, it means each subject shares the same number of the measurement.}

\item{xlab}{a title for x axis.}

\item{legend}{logicals. It defaults to show the legend bar.}
}
\value{
a ggplot object
}
\description{
This function is to plot the distance matrix in dbICC including the segmentation boundary (colored in red) between the within-
distance and between-distance based on \code{ggplot2}.
}
\examples{
# See example for dm2icc

}
\references{
\itemize{
\item Xu, M., Reiss, P. T., and Cribben, I. (2020). Generalized reliability based on distances. Biometrics, to appear. \url{https://arxiv.org/abs/1912.07137}.
}
}
\seealso{
\code{\link{dm2icc}},\code{\link{dm2icc.bt}}
}
\author{
Meng Xu \email{mxu@campus.haifa.ac.il}
}
\keyword{Distance}
\keyword{matrix}
