% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbicc-package.R
\docType{package}
\name{dbicc-package}
\alias{dbicc-package}
\alias{dbicc}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"dbicc")\Sexpr{tools:::Rd_package_title("dbicc")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"dbicc")\Sexpr{tools:::Rd_package_description("dbicc")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"dbicc")\Sexpr{tools:::Rd_package_DESCRIPTION("dbicc")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"dbicc")\Sexpr{tools:::Rd_package_indices("dbicc")} test how to use the pkg
}
\examples{

See example for dm2icc

}
\references{
Reiss et al. (2019). Generalized test-retest reliability based on distances.
}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"dbicc")\Sexpr{tools:::Rd_package_author("dbicc")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"dbicc")\Sexpr{tools:::Rd_package_maintainer("dbicc")}
}
\keyword{package}
