% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2icc.R
\name{dm2icc}
\alias{dm2icc}
\title{Distance-based Intraclass Correlation Coefficient (dbICC)}
\usage{
dm2icc(dmat, nsub, nmea, bootsamp = 1:nsub, adhoc = FALSE)
}
\arguments{
\item{dmat}{A squared distance matrix, of dimension \code{sum(nmea)*sum(nmea)}. Note that the order of the rows or
columns in the distance matrix is sorted by groups of measurements for each subject or individual. The details refer to
Reiss el at., 2019.}

\item{nsub}{Number of the subject or individual.}

\item{nmea}{A vector containing number of the measurement for each subject or individual; 
if \code{nmea} is a scalar, it means each subject shares the same number of the measurement.}

\item{bootsamp}{option to a vector requiring a sample with the replacement. Default 
includes all the elements of the specified subjects or individuals.}

\item{adhoc}{option to a logical variable: whether to apply the ad hoc correction when
estimating the dbICC from a bootstrap sample. Default is \code{FALSE}.}
}
\value{
A scalar, giving the dbICC value
}
\description{
This is a function to calculate the value of dbICC through the related squared distance matrix. 
For example, when the distance among any test-retest datatype can be defined, its reliability 
performed in Reiss et al. (2019) can be obtained by this function. A sample of the certain subjects with 
the replacement, which is used in the boostrap confidence intervals, can be reached as well.
}
\note{
%% ~~further notes~~
}
\examples{
#under construction

}
\references{
\itemize{
\item Reiss et al. (2019). Generalized test-retest reliability based on distances. 
}
}
\seealso{
\code{\link{plotdmat}}
}
\author{
Philip Reiss \email{reiss@stat.haifa.ac.il}, Meng Xu \email{mxu@campus.haifa.ac.il}
}
\keyword{dbICC,}
\keyword{distance}
\keyword{matrix,}
\keyword{reliability}
