% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2icc.R
\name{dm2icc}
\alias{dm2icc}
\title{Distance-based Intraclass Correlation Coefficient (dbICC)}
\usage{
dm2icc(dmat, nsub, nmea, bootsamp = 1:nsub, adhoc = FALSE)
}
\arguments{
\item{dmat}{A squared distance matrix, of dimension \code{sum(nmea)*sum(nmea)}. Note that the order of the rows or
columns in the distance matrix is sorted by groups of measurements for each subject or individual. The details refer to
Reiss el at., 2019.}

\item{nsub}{Number of the subject or individual.}

\item{nmea}{A vector containing number of the measurement for each subject or individual; 
if \code{nmea} is a scalar, it means each subject shares the same number of the measurement.}

\item{bootsamp}{option to a vector requiring a sample with the replacement. Default 
includes all the elements of the specified subjects or individuals.}

\item{adhoc}{option to a logical variable: whether to apply the ad hoc correction when
estimating the dbICC from a bootstrap sample. Default is \code{FALSE}.}
}
\value{
A scalar, giving the dbICC value
}
\description{
This is a function to calculate the value of dbICC through the related squared distance matrix. 
For example, when the distance among any test-retest datatype can be defined, its reliability 
performed in Reiss et al. (2019) can be obtained by this function. A sample of the certain subjects with 
the replacement, which is used in the boostrap confidence intervals, can be reached as well.
}
\note{
%% ~~further notes~~
}
\examples{

##Point estimates of dbICC

# Generation function for R^2 points from multi-normal distribution
R2gen<-function(nsub,nmea,m=1,variance,sds=NULL,pt=F){
    if (is.null(sds)==F) set.seed(sds)
    if (length(nmea)==1) nmea<-rep(1,nsub)*nmea
    sig1<-diag(rep(variance,2))
    mu<-c(0,0)
    sig2<-diag(rep(1,2))
    t<-MASS::mvrnorm(nsub,mu,sig2)
    e<-MASS::mvrnorm(sum(nmea),mu,sig1/m)
    p<-matrix(apply(t,2,rep,times=nmea),ncol=2)+e#(I*J)x2
    if (pt==T) return(list(t=t,p=p))
    if (pt==F) return(p)
}

# set the number of the point
I <- 10

# set the number of the measurement for each point
J <- 4

# generate the sample of R^2 points
varl <- 4 # variance of the 2-d normal distribution
pij <- R2gen(I,J,variance=varl)

# calculate the squared distance matrix via Euclidean distance
distmat<-as.matrix(dist(pij)^2)

#plot the squared distance matrix
plotdmat(distmat,I,J)

# dbICC value
dm2icc(distmat,I,J)

##Bootstrap Confident intervals

B <- 500
bicc <- numeric(B)
for (b in 1:B){
    bs <- sort(sample(I, replace = T))
    bicc[b] <- dm2icc(distmat,I,J,bootsamp=bs)
}

quantile(bicc, c(.025, .975))

}
\references{
\itemize{
\item Reiss et al. (2019). Generalized test-retest reliability based on distances. 
}
}
\seealso{
\code{\link{plotdmat}}
}
\author{
Philip Reiss \email{reiss@stat.haifa.ac.il}, Meng Xu \email{mxu@campus.haifa.ac.il}
}
\keyword{dbICC,}
\keyword{distance}
\keyword{matrix,}
\keyword{reliability}
