% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdmat.R
\name{plotdmat}
\alias{plotdmat}
\title{Plot the distance matrix in dbICC}
\usage{
plotdmat(dmat, nsub, nmea, xlab = NULL)
}
\arguments{
\item{dmat}{A squared distance matrix}

\item{nsub}{Number of the subject or individual}

\item{nmea}{A vector of number of the measurement for each subject or individual}

\item{xlab}{The title for the x axis}
}
\value{
A ggplot object
}
\description{
This function is to plot the distance matrix in dbICC including the segmentation boundary (colored in red) between the within-
distance and between-distance based on \code{ggplot2}.
}
\note{
%% ~~further notes~~
}
\examples{
# See example for dm2icc

}
\references{
\itemize{
\item Reiss et al. (2019). Generalized test-retest reliability based on distances. 
}
}
\seealso{
\code{\link{dm2icc}}
}
\author{
Meng Xu \email{mxu@campus.haifa.ac.il}
}
\keyword{}
